Change type of cost, income, limit value fields to decimal

+++ CostIncomeCalculator.api/Dtos/CostDtos/MonthCostDto.cs
@@ -3,6 +3,6 @@ namespace cost_income_calculator.api.Dtos.CostDtos
     public class MonthCostDto
     {
         public string Category { get; set; }
-        public double CostSum { get; set; }
+        public decimal CostSum { get; set; }
     }
 }
