+++ CostIncomeCalculator.api/Dtos/IncomeDtos/IncomeReturnDto.cs
@@ -7,7 +7,7 @@ namespace cost_income_calculator.api.Dtos.IncomeDtos
         public int Id { get; set; }
         public string Category { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public DateTime Date { get; set; }
     }
 }
